+++ source/Tool Geodes/Overrides/GameLocation.cs
@@ -10,14 +10,16 @@ namespace ToolGeodes.Overrides
 {
     public static class MonsterDamageHook
     {
-        public static bool Prefix(GameLocation __instance, Microsoft.Xna.Framework.Rectangle areaOfEffect,ref int minDamage, ref int maxDamage, bool isBomb, ref float knockBackModifier, int addedPrecision, float critChance, float critMultiplier, bool triggerMonsterInvincibleTimer, Farmer who)
+        public static bool Prefix(GameLocation __instance, Microsoft.Xna.Framework.Rectangle areaOfEffect,ref int minDamage, ref int maxDamage, bool isBomb, ref float knockBackModifier, int addedPrecision, ref float critChance, float critMultiplier, bool triggerMonsterInvincibleTimer, Farmer who)
         {
             double mult = 1.0 + who.HasAdornment(ToolType.Weapon, Mod.Config.GEODE_MORE_DAMAGE) * 0.20;
             minDamage = (int)(minDamage * mult);
             maxDamage = (int)(maxDamage * mult);
 
-            mult = 1.0 * who.HasAdornment(ToolType.Weapon, Mod.Config.GEODE_MORE_KNOCKBACK) * 0.50;
+            mult = 1.0 + who.HasAdornment(ToolType.Weapon, Mod.Config.GEODE_MORE_KNOCKBACK) * 0.50;
             knockBackModifier = (float)(knockBackModifier * mult);
+
+            critChance += (float)(who.HasAdornment(ToolType.Weapon, Mod.Config.GEODE_MORE_CRITCHANCE) * 0.05);
 
             return true;
         }