+++ source/~Pathoschild/ContentPatcher/Framework/Conditions/ConditionType.cs
@@ -63,6 +63,9 @@ namespace ContentPatcher.Framework.Conditions
         /// <summary>An event ID the player saw.</summary>
         HasSeenEvent,
 
+        /// <summary>The special items in the player's wallet.</summary>
+        HasWalletItem,
+
         /// <summary>The current player's internal spouse name (if any).</summary>
         Spouse,
 
@@ -74,6 +77,9 @@ namespace ContentPatcher.Framework.Conditions
 
         /// <summary>The current player's relationship status with the character (matching <see cref="StardewValley.FriendshipStatus"/>)</summary>
         Relationship,
+
+        /// <summary>The current player's level for a skill.</summary>
+        SkillLevel,
 
         /****
         ** Magic conditions